**How**

1. Make our output type a cdylib not a dylib.
2. Rename the cffi-provided `initnative_engine` function to
   `wrapped_initnative_engine`, and create `initnative_engine` in Rust,
   which calls the C version. This is required because otherwise the
   linker will strip `initnative_engine` out of the binary.

**Why**

1. This is the supported path in rust. The fact that making a `dylib`
   happened to work before isn't particularly well defined behavior in
   Rust, but was a nice accident.
2. Binary size. When building a cdylib, the linker can strip dead code,
   and perform LTO. Note that LTO is not enabled in this commit, but
   will be a seprate PR. The following size improvements were observed
   in release mode:
   OSX:
    - dylib (before this change): 16M
    - cdylib (with this change): 2.9M

   Linux:
    - dylib (before this change): 66M
    - cdylib (with this change): 51M

3. Performance. This allows the toolchain to be much more aggressive
   with optimisations (particularly when we enable LTO in the future).
   Time to list a large repo (2nd invocation, without daemon):
   OSX:
    - Before: 4m29
    - After: 4m12

   Linux:
    - Before: 4m11
    - After: 4m7

**Drawbacks**

Slows down release mode compilation.
OSX:
 - Before: 323s
 - After: 348s

Linux (on travis, which has somewhat variable performance):
 - Before: 436s
 - After: 556s